Add workaround for possible SBT classpath problem.
Under some circumstances SBT might mess up the class path which will
cause the loading of a configuration via `ConfigFactory.load()` to not
load the correct file or none at all.

To avoid this we explicitly set the class loader to use. This is
especially recommended if we're using `IO` because that might cause the
thread class loader to be used.